D15_FastAPI/requirements.txt
click==7.1.2
colorama==0.4.4
fastapi==0.63.0
flake8==3.9.1
h11==0.12.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.24
Pillow==8.2.0
protobuf==3.15.8
pycodestyle==2.7.0
pydantic==1.8.1
pyflakes==2.3.1
pytesseract==0.3.7
python-dotenv==0.17.0
python-multipart==0.0.5
PyYAML==5.4.1
six==1.15.0
starlette==0.13.6
tesseract==0.1.3
typing-extensions==3.7.4.3
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
